fusion alarmas con eventos
