+++ event_manager/models.py
@@ -29,3 +29,17 @@ class Event(models.Model):
 
     def __unicode__(self):
         return self.pk + ' ' + self.sensor
+
+
+class Alarm(models.Model):
+    event = models.ForeignKey(Event)
+    activation_date = models.DateTimeField('activation date', auto_now_add=True)
+    activated = models.BooleanField('Is activated?', default=True)
+    notified = models.BooleanField('Is notified?', default=False)
+
+    class Meta:
+        verbose_name = 'alarm'
+        verbose_name_plural = 'alarms'
+
+    def __unicode__(self):
+        return self.pk + ' - Event: ' + self.event
